Mediating in repository between remote and local data

+++ app/src/main/java/com/adyen/android/assignment/data/repositories/AstronomyPictureRepositoryImpl.kt
@@ -9,13 +9,17 @@ import javax.inject.Inject
 class AstronomyPictureRepositoryImpl @Inject constructor(
     private val remoteDataSource: AstronomyPictureRemoteDataSource,
     private val localDataSource: AstronomyPictureLocalDataSource,
+    // The DAO can be used instead of the local data source as there is no extra logic needed, but let's illustrate having a local ds.
 ) : AstronomyPictureRepository {
 
     override suspend fun getPictures(refresh: Boolean, count: Int): List<AstronomyPicture> {
-        if (refresh) {
-            localDataSource.deleteAllPictures()
+        val isCacheEmpty = localDataSource.getPictures(count).isEmpty()
+
+        if (isCacheEmpty || refresh) {
+            if (refresh) localDataSource.deleteAllPictures()
             localDataSource.insertPictures(remoteDataSource.getPictures(count))
         }
+
         return localDataSource.getPictures(count)
     }
 }
